+++ apps/omsorgspengerutbetaling-arbeidstaker-soknad/src/app/utils/søknadsdataToApiData/getApiDataFromSøknadsdata.ts
@@ -6,8 +6,8 @@ import { getAttachmentURLBackend } from '../attachmentUtilsAuthToken';
 import { getMedlemskapApiDataFromSøknadsdata } from './getMedlemskapApiDataFromSøknadsdata';
 import { getUtenlansoppholdApiDataFromSøknadsdata } from './getUtenlandsoppholdApiDataFromSøknadsdata';
 import { getArbeidsgivereApiDataFromSøknadsdata } from './getArbeidsgivereApiDataFromSøknadsdata';
-import { getFosterbarnApiDataFromSøknadsdata } from './getFosterbarnApiDataFromSøknadsdata';
 import { getDataBruktTilUtledning } from './getDataBruktTilUtledning';
+import { getDineBarnApiDataFromSøknadsdata } from './getDineBarnApiDataFromSøknadsdata';
 
 const getVedleggApiData = (vedlegg?: Attachment[]): string[] => {
     if (!vedlegg || vedlegg.length === 0) {
@@ -16,21 +16,19 @@ const getVedleggApiData = (vedlegg?: Attachment[]): string[] => {
     return vedlegg.filter(attachmentIsUploadedAndIsValidFileFormat).map(({ url }) => getAttachmentURLBackend(url));
 };
 
-const getArbeidsforholdDokumenter = (situasjon: SituasjonSøknadsdata): string[] => {
+const getArbeidsforholdDokumenter = (situasjon: SituasjonSøknadsdata): Attachment[] => {
     const dokumenter: Attachment[] = [];
-
     Object.values(situasjon).forEach((forhold) => {
         if (forhold.type === 'harHattFraværUtenLønnKonfliktMedArbeidsgiver') {
             dokumenter.push(...forhold.dokumenter);
         }
     });
-
-    return getVedleggApiData(dokumenter);
+    return dokumenter;
 };
 
 export const getApiDataFromSøknadsdata = (søknadsdata: Søknadsdata): SøknadApiData | undefined => {
-    const { id, fosterbarn, situasjon, fravær, legeerklæring, medlemskap } = søknadsdata;
-    if (!id || !fosterbarn || !situasjon || !fravær || !medlemskap || !legeerklæring) {
+    const { id, dineBarn, deltBosted, situasjon, fravær, legeerklæring, medlemskap } = søknadsdata;
+    if (!id || !dineBarn || !situasjon || !fravær || !medlemskap || !legeerklæring) {
         return undefined;
     }
     const språk = 'nb';
@@ -42,11 +40,15 @@ export const getApiDataFromSøknadsdata = (søknadsdata: Søknadsdata): SøknadA
             harForståttRettigheterOgPlikter: søknadsdata.velkommen?.harForståttRettigheterOgPlikter === true,
             harBekreftetOpplysninger: søknadsdata.oppsummering?.harBekreftetOpplysninger === true,
         },
-        fosterbarn: getFosterbarnApiDataFromSøknadsdata(fosterbarn),
+        dineBarn: getDineBarnApiDataFromSøknadsdata(dineBarn),
         arbeidsgivere: getArbeidsgivereApiDataFromSøknadsdata(situasjon, fravær),
         opphold: getUtenlansoppholdApiDataFromSøknadsdata(språk, fravær),
         bosteder: getMedlemskapApiDataFromSøknadsdata(språk, medlemskap),
-        vedlegg: [...getVedleggApiData(legeerklæring?.vedlegg), ...getArbeidsforholdDokumenter(situasjon)],
+        vedlegg: [
+            ...getVedleggApiData(deltBosted?.vedlegg),
+            ...getVedleggApiData(legeerklæring?.vedlegg),
+            ...getVedleggApiData(getArbeidsforholdDokumenter(situasjon)),
+        ],
         dataBruktTilUtledningAnnetData: JSON.stringify(getDataBruktTilUtledning(søknadsdata)),
     };
 };