+++ apps/sif-ettersending/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @navikt/sif-ettersending
 
+## 3.33.6
+
+### Patch Changes
+
+-   Updated dependencies [a739a56]
+    -   @navikt/sif-common-formik-ds@1.9.0
+    -   @navikt/sif-common-core-ds@9.17.9
+    -   @navikt/sif-common-soknad-ds@10.15.8
+
+## 3.33.5
+
+### Patch Changes
+
+-   1972384: Pakkeoppdateringer minor/patch + uuid major
+-   Updated dependencies [1972384]
+    -   @navikt/sif-common-formik-ds@1.8.8
+    -   @navikt/sif-common-soknad-ds@10.15.8
+    -   @navikt/sif-common-core-ds@9.17.9
+    -   @navikt/sif-common-sentry@0.26.6
+    -   @navikt/sif-common-utils@3.46.3
+
 ## 3.33.4
 
 ### Patch Changes