+++ packages/sif-common-formik-ds/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @navikt/sif-common-formik-ds
 
+## 1.9.0
+
+### Minor Changes
+
+-   a739a56: Håndtere lankode for Kosovo på en bedre måte. Refactoring av CountrySelect.
+
+## 1.8.8
+
+### Patch Changes
+
+-   1972384: Pakkeoppdateringer minor/patch + uuid major
+
 ## 1.8.7
 
 ### Patch Changes